Enhance documentation and implement pointer canonicalization utilities
This commit adds comprehensive documentation for the pointer canonicalization and equivalence checking utilities in the UnderApproxAA module. Key functions such as `stripNoopCasts`, `sameConstOffset`, `isZeroGEP`, and `isRoundTripCast` are documented to clarify their purpose, parameters, and expected behavior. Additionally, the EquivDB class is documented to explain its role in managing equivalence classes of pointer values using union-find with congruence closure. These improvements aim to enhance code readability and facilitate better understanding for future developers.

+++ lib/Alias/LotusAA/Engine/InterProceduralPass.cpp
@@ -337,7 +337,7 @@ void LotusAA::computePtsCgIteratively(Module &M, std::vector<Function *> &func_s
       set<Function *> callersNeedingReanalysis = changed_func;
       changed_func.clear();
       
-      for (int i = func_seq.size() - 1; i >= 0; i--) {
+      for (size_t i = func_seq.size() - 1; i >= 0; i--) {
         Function *func = func_seq[i];
         IntraLotusAA *func_result = getPtGraph(func);
